Some initial forays into gatling with Polr
Experimental.scala is the best bet for url testing so far, it uses
a custom feeder.

LongTest.scala is a generic test run ripped from the Gatling docs
it takes a long time to run.

ShortTest.scala is something similar to the original PolrTest.scala.

PolrTest.scala was removed, but can be restored by ripping the example
on wikipedia.